+++ src/clevertap.js
@@ -40,7 +40,8 @@ import NotificationHandler from './modules/notification'
 import { hasWebInboxSettingsInLS, checkAndRegisterWebInboxElements, initializeWebInbox, getInboxMessages, saveInboxMessages } from './modules/web-inbox/helper'
 import { Variable } from './modules/variables/variable'
 import VariableStore from './modules/variables/variableStore'
-import { checkBuilder } from './modules/visualBuilder/pageBuilder'
+import { checkBuilder, addAntiFlicker } from './modules/visualBuilder/pageBuilder'
+import { setServerKey } from './modules/webPushPrompt/prompt'
 
 export default class CleverTap {
   #logger
@@ -509,6 +510,7 @@ export default class CleverTap {
       closeIframe(campaignId, divIdIgnored, this.#session.sessionId)
     }
     api.enableWebPush = (enabled, applicationServerKey) => {
+      setServerKey(applicationServerKey)
       this.notifications._enableWebPush(enabled, applicationServerKey)
     }
     api.tr = (msg) => {
@@ -580,7 +582,10 @@ export default class CleverTap {
   }
 
   // starts here
-  init (accountId, region, targetDomain, token) {
+  init (accountId, region, targetDomain, token, antiFlicker = {}) {
+    if (Object.keys(antiFlicker).length > 0) {
+      addAntiFlicker(antiFlicker)
+    }
     if (this.#onloadcalled === 1) {
       // already initailsed
       return