+++ src/modules/notification.js
@@ -6,6 +6,7 @@ import {
 import {
   urlBase64ToUint8Array
 } from '../util/encoder'
+import { enablePush } from './webPushPrompt/prompt'
 
 export default class NotificationHandler extends Array {
   #oldValues
@@ -35,6 +36,11 @@ export default class NotificationHandler extends Array {
     return 0
   }
 
+  enable (options = {}) {
+    const { swPath } = options
+    enablePush(this.#logger, this.#account, this.#request, swPath)
+  }
+
   _processOldValues () {
     if (this.#oldValues) {
       this.#setUpWebPush(this.#oldValues)
@@ -53,7 +59,7 @@ export default class NotificationHandler extends Array {
     }
   }
 
-  #setUpWebPushNotifications (subscriptionCallback, serviceWorkerPath, apnsWebPushId, apnsServiceUrl) {
+  setUpWebPushNotifications (subscriptionCallback, serviceWorkerPath, apnsWebPushId, apnsServiceUrl) {
     if (navigator.userAgent.indexOf('Chrome') !== -1 || navigator.userAgent.indexOf('Firefox') !== -1) {
       this.#setUpChromeFirefoxNotifications(subscriptionCallback, serviceWorkerPath)
     } else if (navigator.userAgent.indexOf('Safari') !== -1) {
@@ -61,7 +67,7 @@ export default class NotificationHandler extends Array {
     }
   }
 
-  #setApplicationServerKey (applicationServerKey) {
+  setApplicationServerKey (applicationServerKey) {
     this.#fcmPublicKey = applicationServerKey
   }
 
@@ -153,20 +159,27 @@ export default class NotificationHandler extends Array {
             if (typeof subscriptionCallback !== 'undefined' && typeof subscriptionCallback === 'function') {
               subscriptionCallback()
             }
+            const existingBellWrapper = document.getElementById('bell_wrapper')
+            if (existingBellWrapper) {
+              existingBellWrapper.parentNode.removeChild(existingBellWrapper)
+            }
           }).catch((error) => {
-            this.#logger.error('Error subscribing: ' + error)
             // unsubscribe from webpush if error
             serviceWorkerRegistration.pushManager.getSubscription().then((subscription) => {
               if (subscription !== null) {
                 subscription.unsubscribe().then((successful) => {
                   // You've successfully unsubscribed
                   this.#logger.info('Unsubscription successful')
+                  window.clevertap.notifications.push({
+                    skipDialog: true
+                  })
                 }).catch((e) => {
                   // Unsubscription failed
                   this.#logger.error('Error unsubscribing: ' + e)
                 })
               }
             })
+            this.#logger.error('Error subscribing: ' + error)
           })
       }).catch((err) => {
         this.#logger.error('error registering service worker: ' + err)
@@ -282,7 +295,7 @@ export default class NotificationHandler extends Array {
       // handle migrations from other services -> chrome notifications may have already been asked for before
       if (Notification.permission === 'granted') {
         // skip the dialog and register
-        this.#setUpWebPushNotifications(subscriptionCallback, serviceWorkerPath, apnsWebPushId, apnsWebPushServiceUrl)
+        this.setUpWebPushNotifications(subscriptionCallback, serviceWorkerPath, apnsWebPushId, apnsWebPushServiceUrl)
         return
       } else if (Notification.permission === 'denied') {
         // we've lost this profile :'(
@@ -290,7 +303,7 @@ export default class NotificationHandler extends Array {
       }
 
       if (skipDialog) {
-        this.#setUpWebPushNotifications(subscriptionCallback, serviceWorkerPath, apnsWebPushId, apnsWebPushServiceUrl)
+        this.setUpWebPushNotifications(subscriptionCallback, serviceWorkerPath, apnsWebPushId, apnsWebPushServiceUrl)
         return
       }
     }
@@ -384,7 +397,7 @@ export default class NotificationHandler extends Array {
             if (typeof okCallback === 'function') {
               okCallback()
             }
-            this.#setUpWebPushNotifications(subscriptionCallback, serviceWorkerPath, apnsWebPushId, apnsWebPushServiceUrl)
+            this.setUpWebPushNotifications(subscriptionCallback, serviceWorkerPath, apnsWebPushId, apnsWebPushServiceUrl)
           } else {
             if (typeof rejectCallback === 'function') {
               rejectCallback()
@@ -399,7 +412,7 @@ export default class NotificationHandler extends Array {
   _enableWebPush (enabled, applicationServerKey) {
     $ct.webPushEnabled = enabled
     if (applicationServerKey != null) {
-      this.#setApplicationServerKey(applicationServerKey)
+      this.setApplicationServerKey(applicationServerKey)
     }
     if ($ct.webPushEnabled && $ct.notifApi.notifEnabledFromApi) {
       this.#handleNotificationRegistration($ct.notifApi.displayArgs)