+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.10.1] - 8 Oct, 2024
+- Adding a fix for the version check in visual editor
+
+## [1.10.0] - 8 Oct, 2024
+- Adds new api to handle rendering of customized web push prompt
+
+## [1.9.6] - 23 Sept, 2024
+- Shopify support for visual builder
+
+## [1.9.5] - 23 Sept, 2024
+- Elimninates the use for migration service worker.
+
+## [1.9.4] - 18 Sept, 2024
+- Added anti flicker solution for visual editor
+
+## [1.9.3] - 16 Sept, 2024
+- Fixes a bug where multivalue feature was giving error.
+
 ## [1.9.2] - 4 Sept, 2024
 - Sending SDK and account ID via post message